internal: add syncer
