+++ chain/syncer.go
@@ -1,0 +1,20 @@
+package chain
+
+import (
+	"context"
+
+	"go.sia.tech/core/gateway"
+	"go.sia.tech/core/types"
+	"go.sia.tech/coreutils/syncer"
+)
+
+type (
+	Syncer interface {
+		Addr() string
+		BroadcastHeader(h gateway.BlockHeader)
+		BroadcastTransactionSet([]types.Transaction)
+		Close() error
+		Connect(ctx context.Context, addr string) (*syncer.Peer, error)
+		Peers() []*syncer.Peer
+	}
+)
